refactor: simpleSchema updates
