+++ imports/node-app/core-services/catalog/simpleSchemas.js
@@ -11,7 +11,7 @@ import SimpleSchema from "simpl-schema";
  * @property {String} valueType optional
  * @property {String} description optional
  */
-export const Metafield = new SimpleSchema({
+const Metafield = new SimpleSchema({
   key: {
     type: String,
     max: 30,
