+++ lib/api/router/metadata.js
@@ -9,7 +9,7 @@ import { Shops, SellerShops } from "/lib/collections";
  * @param context
  * @example MetaData.init(Router.current());
  */
-export const MetaData = {
+const MetaData = {
   init(context) {
     const { params } = context;
     const product = ReactionProduct.selectedProduct();
